requirements.txt
pymysql==1.1.0
sqlalchemy==1.4.49
tqdm==4.66.1
python-dotenv==1.0.0 